Enrich Varnish logs with GeoIP data The core "app" container now has 2 new essential integrations which are wired into the `Dev` environment (a.k.a. `just debug`): 1. Vector integrated with MaxMind GeoLite2City data 2. Vector fully integrated with Honeycomb
This also changes the `debug` just command to require `team` access so
that we can experiment with the **REAL** behaviour locally. Since we
need the MaxMind data which requires a license key, we might as well go
all in and grab the Honeycomb API key too.

In the `Test` environment, we do not make the GeoIP data available as the
MaxMind license key will not be made public, which is relevant for
GitHub repository forks.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>