Fix purging across all backends (#15)
As we've discovered with Jerod part of
https://github.com/thechangelog/changelog.com/pull/549, purging was only
working for the app backend, assets & feeds did not work correctly.

It turns out that short-circuiting the request was not helpful.

For the feeds backend, the URLs would get re-written, and then `hash`
would get callled, even though we wanted to... purge.

As for the assets backend, only  the `GET` & `HEAD` methods were
allowed... 🤦‍♂️

This fixes it + adds VTC & acceptance tests to make sure that purging
works as expected, across all backends.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>