separate include for news redirects
I tried to recreate the same behavior I was seeing from the changelog.com.vcl and from sending requests to the Fastly CDN (currently www.changelog.com).

In the VCL used at Fastly it was performing a 308 redirect and included the query string in the redirected url.

I added a WithFile() section in main.go to make sure the new included VCL file gets copied into the container along side the default vcl when the container is built.

Signed-off-by: Matt Johnson <m@ttjohnson.com>