Add redirects from Fastly VCL (#19)
* separate include for news redirects

I tried to recreate the same behavior I was seeing from the changelog.com.vcl and from sending requests to the Fastly CDN (currently www.changelog.com).

In the VCL used at Fastly it was performing a 308 redirect and included the query string in the redirected url.

Signed-off-by: Matt Johnson <m@ttjohnson.com>

* adding vtc and hurl tests for news redirects

Signed-off-by: Matt Johnson <m@ttjohnson.com>

* looser hurl redirect location matching to remove hostname from assertion

using contains keeps it simple and avoids trying to use matches which can have unexpected behavior when trying to match query string special characters thinking they relate to regular expression syntax

Signed-off-by: Matt Johnson <m@ttjohnson.com>

* Fix `just test-acceptance-report` inside the container

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* news-redirects additional updates from feedback

- include assertions for http status code being 308 in hurl acceptance tests
- add the rest of the redirects from fastly (changelog.com.vcl) table news_redirects
- defend against header injection attacks

Signed-off-by: Matt Johnson <m@ttjohnson.com>

- Tick one more item off the README

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* moving all pipedream.changelog.com vcl configs in a directory

using WithDirectory() instead of multiple WithFile()
resolves https://github.com/thechangelog/pipely/pull/19#discussion_r2183391569

Signed-off-by: Matt Johnson <m@ttjohnson.com>

* cleanup http status response and test assertions

- added recommended http status text / reason phrase to news-redirect responses
- removed redundant assertion on status code

Signed-off-by: Matt Johnson <m@ttjohnson.com>

* Add acceptance tests for all News mp3 redirects

So that we compare the behaviour of both CDNs. The idea is that they
should be identical by the time we are ready to roll out 1.0

Also add acceptance tests + redirects for:
- http -> https
- www -> non-ww

Also added speedtest benchmarks. Especially useful when running with
`--cloud` 🤫

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

Signed-off-by: Matt Johnson <m@ttjohnson.com>
Signed-off-by: Gerhard Lazu <gerhard@changelog.com>
Co-authored-by: Gerhard Lazu <gerhard@changelog.com>