Send Varnish logs to AWS S3 (#27)
Uses the same format as the existing logs. Resisting to improve this was
hard, but necessary. We are just lifting & shifting in this phase, will
improve in post 1.0 follow-ups.

We are writing the logs to S3 at least (whichever comes first):
- every hour
- when 100KB of logs are stored in memory

For local production, we are using a different S3 region and suffix the
buckets with `-local-pipedream` so that we experience the full
end-to-end setup, but make it twice as difficult to write to the
production buckets (different region & different bucket name).

While this is OK for now, the improvement that I'm itching to do is
replace **ALL** of this with ClickHouse. While S3 makes sense as a data
lake, the CSV / JSON combo are showing their age. A much better approach
would be to leverage something modern like ClickHouse, and send all
requests to it, without making a distinction. This would allow to
replace:

- Logging logic in the CDN
- S3 log stats
- Stats workers
- Honeycomb.io

This is something that I'm much more excited about, than working around
complexities inherent to the current setup. Something to dig into post
1.0, for sure 😉

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>