Add more locations
This brings the total of locations to 16 (up from 10).

We need to serve **a lot** of traffic in Europe, and this distributes it
a bit more among 4 high-traffic locations:
1. Frankfurt
2. Amsterdam
3. Paris
4. London

We are also adding more capacity on the West coast (Seattle & Los
Angeles), East coast (New Jersey) and also one more location in Asia:
Tokyo.

This makes the entire system hum!

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>