Pin DNS resolution in acceptance tests (#37)
Shortly after https://github.com/thechangelog/changelog.com/discussions/546,
when we went back to resolving 50% of the traffic to Fastly (we'll cover this in
Kaizen 21), acceptance tests started failing when cdn.changelog.com &
changelog.com would resolve to Pipedream.

This fixes it - we resolve DNS explicitly.

Also, we no longer need to run acceptance tests on a schedule. We know
that the setup works. We have live traffic & metrics that continuously
validate the behaviour. We still run these in PRs, merges & releases,
it's just the scheduled part that we are dropping.

Lastly, I lined up the local `just` version with what we are running in
Namespace & GitHub.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>