Support MP3 uploads (#39)
This increases the app first_byte_timeout to 5 minutes so that Varnish
has enough time to upload the file to the app, and receive a response in
a timely fashion.

I tested with a 212MB file which took 22s to upload to Varnish, and
another 54s for the backend to process it. While I could set the
first_byte_timeout to a lower value (e.g. 180s), the 300s values gives
plenty of headroom in case one of the components in the chain is slower
than normal.

As I was testing this out, I had to change the way the container was
running locally because requests kept timing out during admin logins 🤷‍♂️

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>