Store MP3s in file cache + HOT & COLD instances (#44)
This is a big change that includes:
- pipedream.changelog.com removal (it's just changelog.com now)
- break down the monolithic default.vcl into smaller fragments (this
  allowed the tests to be refactored, no more VCL duplication!)
- keep objects in cache to prevent re-downloading if-not-modified
- Varnish tuning to match changelog.com `varnishstat -1`
- split instances into HOT (receive the most traffic) & COLD

All this so that we can configure disk cache for storing those MP3
files. Without a disk cache, the memory becomes fragmented due to all
the large files, and other large files can no longer be stored. After a
long enough period, small files can no longer be stored either, which
means that caching stops working altogether.

We are now storing MP3 files on disk, and everything else in memory.
Once we get this deployed and check the stats, the next improvement will
become obvious.

Another important side-effect of this change is a concrete step towards
allowing VCL re-use across different top-level configs which will be
necessary for v2.0 - making it easy for others to deploy this.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>