Add nightly.changelog.com
Reorganise the acceptance tests & check production AFTER we deploy the
new version (otherwise tests will fail).

Add a 204 backend health probe so that we can support backends that
return HTTP 204, otherwise the backend will be marked as "sick".

Add bench-nightly-* so that we can see how it compares - see more
details in the PR description. While at it, add assets benchmarks too,
and use HTTP 1.1 until `oha` works with both `--http2 --host` flags.

FWIW, if we want this to be useful for others, the code should be simpler for
an arbitrary number of backends. Something that will need to improve as
we approach v2.0.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>