Run periodic checks against all regions
This makes sure that all regions behave as they should, no hanging
requests. We retry ONCE in case some of the files are not in cache. This
is part of the check-all command, no extra config required, because we
expect this to be the case sometimes. If the command fails, there is
definitely a problem with that regional instance.

If this passes consistently after a few days, it's OK to close:
- https://github.com/thechangelog/changelog.com/issues/553

Other improvements worth mentioning:
- setup ~/.local cache on Namespace for the tools that we install JIT
- use Namespace.so artifact storage when on those runners
- prevent job stealing via github.run-id

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>